moe: reorganize Xe2 MoE kernel headers into bf16/ and mxfp4_w4a16/ subdirs
Renames src/sycl/kernels/moe/xe20/{moe_kernel,moe_mainloop}.hpp to
src/sycl/kernels/moe/xe20/bf16/, and moves the W4A16 MXFP4 forks from
src/sycl/kernels/moe/xe20_mxfp4_w4a16/ to src/sycl/kernels/moe/xe20/mxfp4_w4a16/.
Updates include paths in the four call sites.

Per review feedback (PR #184): both kernel families belong under the
same xe20/ tree, named by precision rather than living in parallel
top-level directories.
