+++ src/sycl/GroupGemmMxfp4W4A16Xe20LauncherInstance.cpp.in
@@ -9,7 +9,7 @@
 #include "sycl/Utils.h"
 #include "cutlass/gemm/device/gemm_universal_adapter.h"
 #include "cutlass/gemm/group_array_problem_shape.hpp"
-#include "sycl/kernels/moe/xe20_mxfp4_w4a16/moe_mxfp4_w4a16_kernel.hpp"
+#include "sycl/kernels/moe/xe20/mxfp4_w4a16/moe_kernel.hpp"
 
 using namespace cute;
