+++ src/sycl/GroupGemmXe20.cpp
@@ -9,7 +9,7 @@
 #include "Utils.h"
 #include "cutlass/gemm/device/gemm_universal_adapter.h"
 #include "cutlass/gemm/group_array_problem_shape.hpp"
-#include "kernels/moe/xe20/moe_kernel.hpp"
+#include "kernels/moe/xe20/bf16/moe_kernel.hpp"
 
 using namespace cute;
 using namespace MoE;
